refactor: consolidate download ad into single lib module
