+++ app/components/DownloadAdDialog.tsx
@@ -1,5 +1,5 @@
 import { html, raw } from 'hono/html'
-import { downloadAdDialogMarkup } from '../lib/downloadAdMarkup'
+import { downloadAdDialogMarkup } from '../lib/downloadAd'
 
 /** Pre-rendered download popover with literal AdSense markup. */
 export default function DownloadAdDialog() {
