--- app/lib/remarkDownloadAdPopup.test.ts
@@ -1,107 +1,0 @@
-import type { Root } from 'mdast'
-import { describe, expect, it } from 'vitest'
-import { DOWNLOAD_AD_POPUP_FRONTMATTER_KEY, remarkDownloadAdPopup } from './remarkDownloadAdPopup'
-
-function applyPlugin(tree: Root): Root {
-  remarkDownloadAdPopup()(tree)
-  return tree
-}
-
-describe('remarkDownloadAdPopup', () => {
-  it('adds downloadAdPopup: true to frontmatter when a download link is present', () => {
-    const tree: Root = {
-      type: 'root',
-      children: [
-        {
-          type: 'yaml',
-          value: 'title: Download test\npublished: 2026-06-20\n',
-        },
-        {
-          type: 'paragraph',
-          children: [
-            {
-              type: 'link',
-              url: 'https://example.com/file.zip',
-              children: [{ type: 'text', value: 'ファイルをダウンロード' }],
-            },
-          ],
-        },
-      ],
-    }
-
-    applyPlugin(tree)
-
-    expect(tree.children[0]).toMatchObject({ type: 'yaml' })
-    expect((tree.children[0] as { value: string }).value).toContain(`${DOWNLOAD_AD_POPUP_FRONTMATTER_KEY}: true`)
-  })
-
-  it('throws when a download link is present but YAML frontmatter is missing', () => {
-    const tree: Root = {
-      type: 'root',
-      children: [
-        {
-          type: 'paragraph',
-          children: [
-            {
-              type: 'link',
-              url: 'https://example.com/file.zip',
-              children: [{ type: 'text', value: 'ファイルをダウンロード' }],
-            },
-          ],
-        },
-      ],
-    }
-
-    expect(() => applyPlugin(tree)).toThrow('MDX pages with download links must include YAML frontmatter')
-  })
-
-  it('does not modify frontmatter when link text contains ダウンロード but href has no file extension', () => {
-    const tree: Root = {
-      type: 'root',
-      children: [
-        {
-          type: 'yaml',
-          value: 'title: Regular article\npublished: 2026-06-20\n',
-        },
-        {
-          type: 'paragraph',
-          children: [
-            {
-              type: 'link',
-              url: '/download-page',
-              children: [{ type: 'text', value: 'ダウンロードページへ' }],
-            },
-          ],
-        },
-      ],
-    }
-
-    applyPlugin(tree)
-    expect((tree.children[0] as { value: string }).value).not.toContain(`${DOWNLOAD_AD_POPUP_FRONTMATTER_KEY}:`)
-  })
-
-  it('does not modify frontmatter when no download link is present', () => {
-    const tree: Root = {
-      type: 'root',
-      children: [
-        {
-          type: 'yaml',
-          value: 'title: Regular article\npublished: 2026-06-20\n',
-        },
-        {
-          type: 'paragraph',
-          children: [
-            {
-              type: 'link',
-              url: 'https://example.com/file.zip',
-              children: [{ type: 'text', value: 'ファイルを開く' }],
-            },
-          ],
-        },
-      ],
-    }
-
-    applyPlugin(tree)
-    expect((tree.children[0] as { value: string }).value).not.toContain(`${DOWNLOAD_AD_POPUP_FRONTMATTER_KEY}:`)
-  })
-})
