+++ vite.config.ts
@@ -6,8 +6,7 @@ import mdx from '@mdx-js/rollup'
 import honox from 'honox/vite'
 import rehypeExternalLinks from 'rehype-external-links'
 import rehypeSlug from 'rehype-slug'
-import { rehypeDownloadLinks } from './app/lib/rehypeDownloadLinks'
-import { remarkDownloadAdPopup } from './app/lib/remarkDownloadAdPopup'
+import { rehypeDownloadLinks, remarkDownloadAdPopup } from './app/lib/downloadAd'
 import remarkFrontmatter from 'remark-frontmatter'
 import remarkMath from 'remark-math'
 import remarkMdxFrontmatter from 'remark-mdx-frontmatter'
